+++ src/dblite/__init__.py
@@ -5,11 +5,11 @@ Released under the MIT License.
 
 @author      Erki Suurjaak
 @created     18.11.2022
-@modified    07.12.2022
+@modified    27.03.2023
 ------------------------------------------------------------------------------
 """
-__version__      = "1.2.0"
-__version_info__ = (1, 2, 0)
-__version_date__ = "2022-12-07"
+__version__      = "1.3.0.dev2"
+__version_info__ = (1, 3, 0, "dev3")
+__version_date__ = "2023-03-27"
 
 from . api import *